+++ gatsby-config.js
@@ -13,6 +13,7 @@ module.exports = {
     "gatsby-plugin-postcss",
     "gatsby-plugin-image",
     "gatsby-plugin-sitemap",
+    "gatsby-plugin-meta-redirect",
     {
       resolve: "gatsby-plugin-manifest",
       options: {